Prevent fixImports() from erroring if it cannot find rules
If the definition for a one of the rules that we are looking into cannot
be found (e.g. if the eslint-plugin-react package is not installed),
then the rule's message will be like:

> Definition for rule 'react/jsx-no-undef' was not found

When this happened, our regex didn't match and then we ended up with an
error. To avoid this, I am adding a check to ensure that we skip over
these rules.