Remove example-project
I haven't used this in a while. I have a different example project on my
machine that I use instead. I think it might make more sense for this to
be its own repo, but for now I am just jettisoning it because it is just
cluttering up this repo.