Remove message when variable is missing
Plugins should themselves guarantee that a variable name is passed in.
For plugins that use the command line, a missing/empty word wouldn't
even reach this code, because

  importjs word path/to/file.js

(i.e. missing word) will lead to a different error:

  error: missing required argument `pathToFile'

I checked the Atom plugin, and it looks like we return early if there is
no word. So we're all good.