Move message about successful import
I'm working to improve messaging. One goal I have is to move all
messaging into the import/fixImports/addImports/goto methods, and have
them display a single message to summarize the results. This commit
moves the message displayed when a variable can is successfully
imported.

While I was at it, I simplified the code a little.