Move anyNumbersWithinRange off of Importer class
This is really a standalone function that does not need to be part of
this class. I thought about putting it in its own module, but I decided
to keep it here for the time being to keep it local. We may want to move
it to its own module to test it directly at some point though.