Add messaging for `fixImports`
Here, we display the number of added and removed imports combined into
one compact(ish) message. If there's only one added/removed, we display
the name of the variable.

While doing this, I spotted a bug with how we consider an unused
variable to be part of imports or not. If there are no previous imports,
and the first line of the file is an unused variable (but not an
import), we would treat it as a removed import. We wouldn't delete the
line, but we would include it in the message. I found this on a test
having this as `text`:

  var a = <span/>;

Without the fix, the messaging would end up including this variable as a
removed import, e.g. "Removed `a`.".