Don't ask to resolve variables with no matching modules
I noticed a bug when trying to import a variable that there's no
matching module for. For some reason, a prompt asking me to resolve the
import popped up (I'm using the vim plugin). The list of choices are
empty however.

This was because we weren't checking the size of the returned list of
jsModules before marking it as unresolved.