Add babel-plugin-add-module-exports
This will make it so we can use `export default` style syntax without
the pain of having to specify `.default` when using require, or
switching from require to import. We might eventually want to switch to
use import for lib code, but I imagine that we will continue to use
require in tests, so this seems like a good long-term solution for us.