Don't show "Removed" message for non-imports
This ended up being more complicated than I thought. The main issue I
started out solving was that a file looking like this

  const foo = <span/>;

(...and nothing else).

Would have the message "Removed `foo`" when running import, despite
no changes having been made to the import. My initial fix for this in
8289e7f (ignoring ranges that were of length 0) had the side-effect of
not removing unused imports if the file looked something like this:

  import foo from 'foo';

(...and nothing else).

The main problem was that the lodash.range we were using could only
communicate where the block of imports started and ended, but not if it
was completely empty. The two examples above would both have the
following range:

  [0]

By instead using a plain object with `start` and `end`, we can correctly
communicate where imports start and end, and also if the block is empty.
For the two examples above, we now have the following ranges:

  { start: 1, end: 2 } // one import
  { start: 1, end: 1 } // empty

I'm continuing to be confused by the one-indexed line numbering, so I'm
going to take a stab at fixing that next.