Extract findMatchingFiles and formattedToRegex into own modules
This is in preparation for mocking the file system in the Importer test
(which should speed things up significantly).

I added a basic test for findMatchingFiles. formattedToRegex is
implicitly tested via that test, and I don't know if it makes sense to
test individually.

While I was at it, I updated a comment for formattedToRegex. I also
fixed a bug where the find command would return an empty string that we
would process.