Use camelCase for configuration instead of snake_case
Now that we are entirely in a JS world, it feels really foreign to have
stuff in snake_case. I think it will be good for us to switch over to
camelCase before we release 1.0.0.

While I was doing this, I noticed that we have at least one test that
sets a configuration option that does not exist (lookupPath instead of
lookupPaths), which I plan on fixing shortly. I think this would be a
good time to add some basic configuration validation so I'll probably do
that too.

Fixes #243