Set lookupPaths instead of lookupPath in an Importer test
I noticed that this test was setting a non-existent configuration value.
When I set it, the output changed so I simply updated the test to match
the new output. I think this might resolve the TODO we had left earlier
about there maybe being a bug in there, but I didn't have enough context
to actually tell.

I actually found it a little difficult to determine what configuration
was set for this test because it was spread out through various nested
blocks. I think when we go through to make these tests faster, we should
probably also make an effort to flatten some of these things and move
more of the test setup to be more localized to make them easier to read.