Move path finding hack to bin/importjs.js
452071772 ended up breaking a fix I had done in f24b4f26 to make eslint
find plugins and extends relative to cwd as well (instead of only
relative inside import-js). I'm not 100% sure why it broke, but moving
the hack back into the unprocessed bin/importjs.js file fixes the
problem.

We should probably have an integration test to better catch this. For
now we're saved by the fact that I use the tool every day on a code base
with local eslint plugins.