Speed up Importer test by moving `require`s to the top
I was tracking down bad performance of the Importer test and found that
we spent a lot of time in `require`. By adding a `__reset` method to the
FileUtils mock, I was able to move inline `require`s to the top, cutting
the test time in half (down from ~40s to ~20s for me).