Use default config in Importer test
Previously, tests depended on a lookup path of 'app', and a current file
named 'test.js' (conceptually located in the same folder as the 'app'
folder). This resulted in imports not using relative paths, because
test.js wasn't in the same lookup path as the files. To make the test
more like what people will experience on a fresh import-js install, I'm
moving over to using '.' as the lookup path (which is the default). The
result is that most tests now generate relative import paths.