Fix using absolute paths for path to current file
We had code in place which would strip out the path to the current
directory from the path to the current file. That bit of code wasn't
working, since we were feeding the String.replace with a string instead
of a regexp.

This resulted in relative paths not working correctly in Sublime (which
passes in absolute paths for pathToCurrentFile). We could make that
plugin also pass local paths, but I figured it was worth fixing this in
the main project anyway, because the command-line tool needs to support
these paths.

Fixes #263, which is likely a duplicate of #233.