Replace babel-preset-airbnb with babel-preset-es2015
The Airbnb preset recently switch from babel-preset-es2015 to a version
that does not add use strict to every file. It also contains some things
that we don't really need here since we are working in a node
environment, so I decided to simplify this a bit and go for the es2015
preset instead.