Replace eslint-config-airbnb with eslint-config-airbnb-base
This package has been split up into a base package for regular
JavaScript, and another package that includes everything, including
React rules. Since we don't actually use React or JSX in this project,
we can switch to the base package and simplify a little.

One downside is that we have a few tests that depend on the React plugin
being configured here, so I've kept that around. As I mention in a
comment, this is weird and it would be nice to scope this to just the
tests that depend on it, but this will do for now.