Add Flow
I'd like to use a static type checker to make our code mode readable and
to help us find bugs before they happen. It seems that our options are
currently Flow or TypeScript. I don't have experience with either
really, so I decided to give Flow a whirl. Since we are already partly
in the Facebook ecosystem with Jest, it seems like this might be a good
bet.

This adds some flow annotations to a few of our files, to see how it
feels. It ended up finding one possible bug already, which is a nice
sign. If we like the direction this is heading, we can expand it to all
of our files, and then eventually remove the `// @flow` pragma and just
have it on automatically.

One somewhat unfortunate downside I ran into was the need to add a Babel
transform for class properties, since Flow requires annotations on class
properties and this is the way it achieves that. This is only
unfortunate because this hasn't been solidified enough in the spec to
make it to stage 3, so it may end up changing dramatically or never
making it into the language officially, which might mean that we'll end
up with some tech debt here. However, I think the tradeoff is worth it
in this case and since we are already going in on a non-standard Flow
syntax, it seems like we might as well do this too.