Move mocking after importing in tests
Now that we've adjusted the way our tests are set up, we can move these
mock calls after we import the modules. This is better because imports
are hoisted so it is best to keep them at the top if possible to avoid
confusing behavior. It also makes these files work better with
import-js.