Move checkCurrentVersion off of Configuration class
This simple function is not part of Configuration's public API, so it
doesn't really need to be on the class. Moving it off makes our code a
little more functional, which I think will help us move things around if
we want to, make some things easier to test, among other benefits.