Use relative paths for named exports
We want to import named export paths the same way we import regular
paths. If you are using relative paths, we should use relative paths for
named exports as well. In the past, I've made an assumption that named
exports are mostly for package dependencies, and I've sort of ignored
that they can point to regular files.

To add support for relative paths, we need to do a couple of things:
- Use `require.resolve` to find out if the "real" path is a package
  dependency or not. If it is a package dependency, we shouldn't use
  a relative path.
- We also need `require.resolve` to be able to use the fromFile argument
  when asking for configuration.
- Right now, relative paths are only used when the lookupPath is the
  same as the file you are importing to. Since we don't have a
  lookupPath in this scenario (we can check config, but that can
  potentially give us an array back), we assume '.' as the lookupPath.
  This is a bit of a hack, but it does the trick. I'm hoping we can move
  away from lookupPaths soon anyway.

I changed the requireResolve mock a little so that it would be easier to
inject more paths as you go along.

With this change, a bunch of tests had to be updated to specifically
point out package dependencies. Otherwise they would have ended up being
relatively resolved and in the wrong import group.

Fixes #270