Add Node based file finder and use on Windows
The default file finder (findMatchingFiles) executes a `find ... | egrep
...` command to find files matching a variable name. This doesn't work
on Windows, so I'm exploring a native Node file finder as an
alternative.

I'm using glob [1], which is designed to work cross-platform.

From what I can tell, this is the only big thing to take care of in
order to make import-js work on Windows. I'm currently running tests on
the sublime plugin to make things work cross platforms, and my initial
tests are positive.

[1]: https://github.com/isaacs/node-glob