Fix ignoring node_modules in findMatchingFilesWithNode
The previous ignore pattern didn't work. Due to how we use a temp folder
as the lookup path in our findMatchingFiles test, it's tricky to add a
test for ignoring the node_modules folder. There's room for improvement:

 - we need a test
 - we should probably ignore nested node_modules folders as well

If I have time, I'll come back to this soon to make it more robust.