Add `importjs watch` command to speed up finding files
The main performance bottleneck at the moment is the process of finding
matching files for a variable. We use two different strategies at the
moment, both being quite small on larger code bases:

- Shell out a `find` + `egrep` command
- Use the `glob` npm module

In this commit, I'm exploring a third option: keeping an updated list of
files (an index) using Watchman [1]. Assuming that you have watchman
installed, you start a watch process by running `importjs watch`. This
will hook up a subscription to Watchman and keep an updated list of
files stored in `.importjs.index.json`. That way, finding files will be
a lot faster, since now we're only iterating over an array in memory.

This is a naive first approach. We probably have to follow up with some
of these changes:
- Avoid race conditions stemming from the fact that multiple threads
  will want to write to the index file at the same time.
- Removing the index file when the `importjs watch` process ends (so
  that we don't keep using an outdated index).

I'm going to try out this idea for a while and see how it works. The
idea is that editor plugins can setup a watch on startup (or potentially
lazily on first import), and then use the index file while operating.
It's also possible that we'll go down the path of having a running
importjs process accepting "commands" on stdin. That way we wouldn't
have to write the index to disk.

[1]: https://facebook.github.io/watchman/