Make file finding async
I've been wanting to explore a watchman-powered file finder. Before I
can do that, I want all file-finding to be async. This is good, because
we can parallelize lookups.

Instead of adding callbacks all over the place, I decided to use
Promises. I believe this will make the code a little simpler to
understand as it provides a nice interface for everything async.

An annoying issue I ran into while switching to Promises was that I
wasn't getting errors surfaced when they happened. It turns out I was
bit by the auto-swallowing of errors that Promises have. I came across
this article explaining things:
http://jamesknelson.com/are-es6-promises-swallowing-your-errors/

In that article, a third-party solution to globally fix the issue is
suggested. I decided against bringing in another dependency, and instead
bite the bullet and go down the explicit (and somewhat verbose) path,
using `reject` and `catch`.

To make tests easier to work with, I decided to make the asyncness
synchronous by adding a dependency to deasync [1]. I found that from
this SO thread:
http://stackoverflow.com/questions/16586640/node-js-async-to-sync

In the long run, we might want to avoid having that dependency and
instead make test assertions async as well. But for now it allowed me to
move forward without having to rewrite the (quite massive) test suite
that we have.

[1]: https://github.com/abbr/deasync