Add simple daemon
It's not useful yet, but I'm using it as a POC that I can write to
STDIN, react to input, then write to STDOUT. The current version simply
prints the same output you provide.

Here's how to test it in mac osx:

Create a pipe that we can use to communicate with:
 > mkfifo IMPORTJS

Start the daemon using the pipe we created as STDIN:
 > importjsd < IMPORTJS

Write to the pipe:
 > echo 'galooshi' > IMPORTJS

Watch for 'galooshi' to get written to the console where the daemon is
running.