Make watchman file finder honor the `excludes` config option
The watchman file finder is starting to stabilize, and I really need to
add tests for it. I found a bug while testing it out locally. The
`excludes` config option is meant to filter out files, but that wasn't
working for the new finder because we didn't normalize the lookup path
the same way the file paths were.