Redirect console logging to file
We need to be careful with what we're logging to stdout, because clients
will be depending on json content in a certain format ending up there.

We log to a file called `importjs.log` in the os's default tmp
directory. Since this folder isn't super easy to discover, I'll follow
up by logging the path to the logfile on startup.