Exit daemon if dead parent detected
By having all editor plugins pass in their process pids, we can make
sure that the importjsd process never turns into a zombie.

There's a chance that the original parent has died, but a new process
has taken over the PID of the old parent. In that case, the importjsd
process will become a zombie, even with the code included in this
change. That's a somewhat unlikely scenario though, that I don't think
we need to bother much about. According to [1], PIDs rotate (up to
32768) on linux systems so there would have to be a lot of processes
started before the old PID is reused.

Also included in this commit is changing the first log of the command to
be one line. This is important for editors so that they know when they
can start expecting JSON output on the channel.

[1]: http://stackoverflow.com/questions/11323410/linux-pid-recycling