Rename regexp variables in findMatchingFiles.js
There was some confusion around these variables, so I made it clearer
what is a compiled regexp and what is just a string pattern.

Before making the rename, I went down the path of switching to passing
in a compiled RegExp to all places. But I couldn't get
the `egrep` call inside `findMatchingFilesWithFind` to work with the
`toString()` output of RegExp. So I decided to keep the string
throughout all finders (so that it's consistent).