Make WatchmanFileCache work with dynamic `cwd`
This is a follow-up to the work I started to make it possible to use
a single ImportJS instance across multiple working directories. This is
a prerequisite for the Atom plugin, where multiple projects with
different root folders can be open at the same time. My previous commit
made the `workingDirectory` a part of the constructor for the `Importer`
class, passed along for anyone needing it. This commit fixes the
previous singleton WatchmanFileCache module to instead be a class that
you instantiate. You're meant to use the `getForWorkingDirectory`
factory method. It will cache previous instances and return the one
active for the current working directory.