Use better temp dir for `findMatchingFiles` test
Now that we have the ability to specify the working directory (cwd), we
no longer have to keep the temp directory that aid the findMatchingFiles
test in the project root. Instead we can create a folder in the official
os-specific temp directory. This is good because we get better test
coverage of the "dynamic working dir" support.