Move node path fixing to where it is needed
While developing the atom-import-js plugin, I noticed that fixImports
wouldn't work for a project of mine which is using custom eslint
plugins. When the CLI tool is used, this was taken care of in
fix-node-path.js where the NODE_PATH env variable was modified
dynamically. In the Atom plugin, we use the javascript api directly and
we weren't using the node_path fix, which we actually can't do because
we need to use import-js with (potentially) multiple working
directories.

I moved the fixes from fix-node-path.js into a new
getEslintWithPathCorrected.js file, a module exporting a single function
that we then call from fixImports when needing access to `eslint`. I've
tested that this works manually. There's quite a setup to be made in
order to automatically test it, so I'm leaving this without unit tests
(except for smoke-testing the code).

This isn't actually a final fix for the eslint path issues. The path fix
only works once. When the eslint module has been imported for a working
directory, it won't work for another one. I've tried deleting the
require cache for the module in between calls, but that doesn't seem to
help. I'll have to explore further solutions to that. This fix is still
probably good enough, since the bug will only happen to people using
Atom, having two projects using import-js, both with custom plugins that
the other project(s) don't use.