Use in-publish to ensure prepublish is only run at right time
Without this, the prepublish scripts are run when `npm install` is run.
I've run into some situations where this causes `npm install` to not be
able to run during development without removing the prepublish scripts.
To prevent this pain going forward, I am bringing in the in-publish
package to make sure that these are only run when running `npm publish`,
which is what we intended anyway.

The not-in-publish part is there to prevent the script from failing when
not in publish.