Add support for javascript config files (.importjs.js)
This is a step in the direction of allowing a more elaborate
configuration setup, where some configuration options can define a
function to return a value instead of just specifying a value:

  declarationKeyword: (fromFile) => {
    if (/.*\-test\.js$/.test(fromFile)) {
      return 'const';
    }
    return 'import';
  }

This commit DOES NOT implement support for the above ^^. Instead, it
simply allows you to store your configuration in a javascript
module/file. Later commits will take care of making it more dynamic.

For now, we simply `require` the file. I've considered using some of the
packages that eslint uses [1] to load its configuration, but they seem a
little overkill now, at least before I know if the things they solve are
actual problems.

I decided to move over our own config to a js file so that we're living
in the future now.

[1]: https://github.com/eslint/eslint/blob/e0d4b19028addac2413e30b94bd045b10e45cedb/lib/config/config-file.js#L158