Warn users if the js config file doesn't export anything
A mistake I made when converting the json config file to a javascript
file for import-js was forgetting to add a `module.exports = ` in the
file. This led to hard-to-track-down problems, because an empty object
was used instead of my elaborate configuration object. An empty object
was used because that's what node's `require` function returns for any
module that doesn't export anything.