Temporarily disable ESLint no-shadow rule
Very recently, this project started failing linting in CI because of the
no-shadow rule. It seems to now erroneously pick up flow type
definitions, causing a bunch of variables to be flagged as shadowing a
variable.

I don't have time to figure out exactly what is going on right now, so
in the meantime I am just disabling this rule to keep things running
smoothly.