Move env core modules to configuration (#290)
This will allow environments to specify more things than just core
modules. We need this to fully implement Meteor support (epic story
covered in #289).

Since you can specify multiple environments to inherit from, we need to
merge some configuration options. Most importantly, we need to merge
coreModules. We probably also want to merge aliases and namedModules.
Even though we currently only use coreModules in our environment
configs, I decided to add support for merging aliases and namedModules
as well.

This is a little bit more complicated than it needs to be. That's mostly
because we're still supporting local configuration. Once we move off of
that, we can simplify things.

Fixes #290.