Move packageDependencies to configuration
This will allow the meteor environment to implement its own logic to
resolve package dependencies.

The findPackageDependencies function needs to know the current working
directory and whether to include devDependencies or not. In order to
make those values available to the function in the configuration object,
I decided to pass along the `Configuration` instance itself. This is a
bit awkward, but gets the job done. To make this happen, I had to make
the anonymous `mergedValue` function an instance method again (which
essentially reverts bd160be).