Recognize 'meteor/somepackage as a package (#305)
In importStatemetnPathType, the lines

  const path = importStatement.path.replace(/^(.*?)\/.*$/, '$1');
  if (packageDependencies.indexOf(path) !== -1) {

strip a meteor package path such as 'meteor/aldeed:simple-schema' down to nothing but 'meteor' and then fail to match the package name contained in the packageDependencies array.

The regex line was placed there to allow imports for modules inside packages to be properly placed in the package group.

Rather than change the regex to skip the 'meteor/' prefix, I have chosen in this change to eliminate the need for the regex by matching the packageDepencies against either the whole path or the start of the path up to a path divider as shown below.

  if (packageDependencies.some((pkg: string): boolean =>
        importStatement.path === pkg ||
        importStatement.path.startsWith(`${pkg}/`))) {

This should fix the problem in a fashion that is not Meteor specific.