Add `moduleSpecifier` configuration option
This new option will allow people to manipulate the resulting module
specifier. You can for instance use it to add a prefix to the specifier,
e.g.

  moduleSpecifier: ({ moduleSpecifier } => '!lazy!' + moduleSpecifier);

Will result in something like

  import foo from '!lazy!foo';

It's probably mostly useful when the specifier is manipulated
conditionally, either based on the original specifier, the file being
imported to, or both.

  moduleSpecifier: ({ moduleSpecifier, pathToCurrentFile } => {
    if (/-test/.test(pathToCurrentFile)) {
      // Import a mocked version in test files
      return 'mocks/' + moduleSpecifier;
    }

    if (/^meteor/.test(moduleSpecifier)) {
      // Add a leading slash to meteor packages
      return '/' + moduleSpecifier;
    }

    // Fall back to the original specifier
    return moduleSpecifier;
  },