Extract Meteor 3rd-party package exports
This implements a "90%" solution to automatically creating namedExports for the 3rd-party Meteor packages used by an app. It works for the majority of the packages used today, primarily those still using "api.exports" calls in the package.js file to define their interface to users. Those that have already converted to the new api.mainModule calls and do not also use the old api.exports interface are not yet supported.