Add test for meteorPackageNamedExports
Added a test that verifies that meteorPackageNamedExports can extract namedExports from the metadata of either a 3rd party or a local package.

Fixed a problem with specifying the isopackRoot when processing local package metadata that was identified by the test.

Eliminated redundant fs.existsSync checks of whether files exist prior to utilizing readJsonFile.