Configure Travis to push code coverage to CodeClimate
We want this project to be resilient and we believe that code coverage
will help us identify gaps in our testing.

The CodeClimate documentation is a little unclear about how to properly
set this up, so I am trying this out. I believe that I also need to set
the CODECLIMATE_REPO_TOKEN environment variable, which I have added
through the Travis interface to prevent this secret from being checked
into code.

  https://travis-ci.org/Galooshi/import-js/settings

I've manually uploaded current code coverage to CodeClimate and it looks
like we are at 93%.

  https://codeclimate.com/github/Galooshi/import-js/coverage

I've also configured CodeClimate to update Pull Requests with an
integration.