Log process.cwd() when starting daemon
I have noticed that some paths are behaving a bit strangely, and have
noticed a difference between running importjs directly and via the Vim
plugin. To help trace some of these issues, I wanted to know what
directory was being wired up when Vim starts, and I figured this might
be generally useful information so I added it to the log.