Prevent normalizePath from skipping dot-files
I was looking into an issue and logging a bunch of normalized paths, and
happened to notice that files starting with a dot were not being
correctly normalized to "./.file". So I decided to fix this and write
some tests.

I'm not really sure if this was causing any real-world issues, but it
likely was.