Remove stripFromPath in favor of moduleNameFormatter
This option is mutating paths in between configuration calls. This is no
good because it means that configuration cannot rely on paths being
consistent when performing checks (e.g. when trying to determine if
something is a test file).

I originally started working to move this so it doesn't mutate the path
in JsModule at all. After I had that working, I realized that we
recently added a new configuration option that could be used to solve
the same use-case as this one: moduleNameFormatter. Since this is
currently causing bugs and since we already have a replacement ready, I
decided to simply remove this. This is a breaking change.

We still have some other options that end up mutating paths between
configuration checks (e.g. useRelativePaths), which we should try to
resolve. But in the meantime, this is progress. We should add some tests
that verify that paths don't change between configuration checks,
regardless of what they are returning.

Addresses #327