Move config.get call out of possibly tight loop
I have noticed that the path to the current file that is passed to
configuration calls starts with "./" whereas the path to the imported
module does not. I'm tracing down why this discrepancy exists and I
spotted this call to `config.get` that is in a tight loop. Since this is
likely not a super trivial function call and it won't change over the
course of the loop, I decided to move it up outside of the loops. I
expect this to have some small performance improvement.