Ignore tests and mocks when building
These are processed by babel on the fly, thanks to our babel-jest
plugin. We really don't need to be shipping the built versions of these
files either. Since we build the project whenever running `npm link`,
and I'm running that a lot for testing reasons, I've decided to ignore
these files to speed that up.