Remove lookupPaths option
I have noticed that in configuration functions, `pathToCurrentFile`
stats with "./" and `pathToImportedModule` doesn't. It is important for
these paths to be consistent and stable, so I am trying to get to the
bottom of this issue.

I believe that I have traced this back to a line in `JsModule`:

```js
 jsModule.filePath = normalizePath(relativeFilePath);
```

`normalizePath` here simply strips the leading `./` or `.` from the path
if it exists. I believe that this is in there in order to support the
`lookupPaths` option--otherwise, we could just leave the `./` in there
and every path would be relative to the project root here, which
simplifies things.

Taking a step back, `lookupPaths` causes confusion, and it seems like
the desired functionality can be solved with `moduleNameFormatter`. I
would like to remove this option to simplify the API and our code.

I may be wrong, but I think that `useRelativePaths` can also be
accomplished by `moduleNameFormatter`, so I'd also like to remove that
option at another time.

This commit gets us most of the way to removing lookupPaths. I ended up
hard-coding some `node_modules/` stripping in a few places which I'd
like to revisit at some point. Also, there are a few failing tests for
the meteor environment that I wasn't really sure how to appropriately
resolve, so I left them as is for now. I hope to circle back with
@rhettlivingston to figure out how to best handle these tests.

This is another breaking change.