Leave leading dot on paths in JsModule
I have noticed that in configuration functions, `pathToCurrentFile`
stats with "./" and `pathToImportedModule` doesn't. It is important for
these paths to be consistent and stable, so I am trying to get to the
bottom of this issue.

I believe that I have traced this back to a line in `JsModule`:

```js
 jsModule.filePath = normalizePath(relativeFilePath);
```

`normalizePath` here simply strips the leading `./` or `.` from the path
if it exists. I believe that this was in there in order to support the
`lookupPaths` option--otherwise, we could just leave the `./` in there
and every path would be relative to the project root here, which
simplifies things.

Since I just removed the `lookupPaths` option, I believe we can safely
remove this, which seems to resolve the issue.