Remove useRelativePaths option
Now that we have the moduleNameFormatter option, which receives the path
to the module being imported as well as the already-formatted relative
path, users should be able to replicate this option within that context.
Since this is pretty easily achievable that way, I think we should
remove this option to simplify our code and simplify the choices that
users have to make.