Remove some indirection from Importer test
This test file is a beast. It is difficult to understand what is
happening when a test fails, and it is difficult to know what is being
tested or where to add new examples. To help make this better, I am
removing some indirection by migrating things from `beforeEach` blocks
into `it` blocks. There's still a lot more of this that can be done, but
this is a step in the right direction.