Partial Extract ES6 xface using RegExp
This is an experimental attempt to extract an ES6 module's interface
as described by its 'export' statements using a RegExp based approach.

The experiment has likely failed but is worth keeping in case an idea
strikes.

The extractInterface routine is believed to be pretty robust with the
glaring exception of not yet doing anything with a bindingList.

  export var|let|const <bindingList>;

The difficulty with pulling the bindingIdentifiers out of a bindingList
is that they can have initializers between them. Initializers can hold
expressions and expressions can have commas within them. Skipping these
to find the next bindingIdentifier in the list is non-trivial.

One thought is that it might be possible to use this for everything but
the bindingList and then use a true parser or pieces from one for the
bindingList. That may be worth pursuing if performance targets can't be
reached with a full parser based approach.