Remove unused code in JsModule#resolvedFilePath
We use the resolved file path in the `goto` method, and when we pass in
`pathToImportedModule` to configuration functions. If the file found for
a module is a package.json, we want to use the `main` file when
resolving the "true" file path. When we converted this project from Ruby
to JavaScript, I made a mistake when converting some code handling this.
`startsWith` should have been `endsWith`. We have tests around this, and
they are passing even with the faulty code in place. Which means we are
handling it somewhere else. A quick git-blame led me to 5b9b1031 which
adds code that replaces package.json with the main file:

  if (mainFile) {
    jsModule.filePath = `${importPath}/${mainFile}`;
  }

The fix is to simply remove the faulty code.

[Fixes #335]