Add `importjsd logpath` take 2
My original attempt was reverted by 5c5c3ea7 because it broke the
--parent-pid option. I fixed this by using the commander API a little
better.

This commander package has confusing behavior. Specifically around the
`.command()` function. Apparently if you give it a description argument,
it will treat it like git-style subcommands which means that it looks
for a different file for the action instead of allowing the action to be
defined. And, if you don't specify a description, there isn't a built-in
way to set a command as default via the `isDefault` option. I worked
around this by modifying the arguments that are parsed to set my own
default. Feels a bit hacky, but it works.