First shot at representing ES modules
EsModule and EsModuleModule are the first strokes at creating a facility
for discovering, storing, and acccessing the interface of a module.

EsModule roughly corresponds to the "Abstract Module Record" defined in
the ES7 specification at
https://tc39.github.io/ecma262/#sec-abstract-module-records.

EsModuleModule roughly corresponds to the "source text module record"
concrete subclass of EsModule defined in the ES7 specification at
https://tc39.github.io/ecma262/#sec-source-text-module-records.

A container class titled "EsModules" is in the works.

It is envisioned that we will have other concrete subclasses of EsModule
such as EsModuleCJS, EsModuleMeteor and EsModulePredefined with the main
difference being the content of their constructors.

A goal is that code creating imports from the knowledge stored here
will only have to work with EsModules and EsModule.